experiments/langgraph-multi/requirements.txt
# Same dependencies as the single-agent LangGraph experiment
langgraph==0.2.45
psycopg2-binary==2.9.10
httpx==0.28.1
numpy==2.1.3
python-dotenv==1.0.1
